Renames Route enum to Resource
