+++ src/main/java/server/httpserver/HttpServer.java
@@ -24,8 +24,6 @@ public class HttpServer {
     }
 
     public void processRequest() {
-        System.out.println("Listening for request.....");
-
         ExecutorService executor = executorServiceFactory.create();
         ProcessClientRequestTask task = new ProcessClientRequestTask(serverSocket.accept(), requestParser, httpRouteProcessor);
         ThreadExecutorService requestProcessorThread = new RequestThreadExecutorService(executor);
